Rename *_mapper to *_pipeline files
