+++ src/colmap/controllers/hierarchical_pipeline.h
@@ -29,7 +29,7 @@
 
 #pragma once
 
-#include "colmap/controllers/incremental_mapper.h"
+#include "colmap/controllers/incremental_pipeline.h"
 #include "colmap/scene/reconstruction_manager.h"
 #include "colmap/scene/scene_clustering.h"
 #include "colmap/util/base_controller.h"
