+++ src/colmap/controllers/incremental_pipeline.h
