+++ src/colmap/exe/image.cc
@@ -29,7 +29,7 @@
 
 #include "colmap/exe/image.h"
 
-#include "colmap/controllers/incremental_mapper.h"
+#include "colmap/controllers/incremental_pipeline.h"
 #include "colmap/controllers/option_manager.h"
 #include "colmap/image/undistortion.h"
 #include "colmap/scene/reconstruction.h"
